+++ NEWS.md
@@ -1,3 +1,5 @@
+# treasury (development version)
+
 # treasury 0.3.0
 
 * Migration to data.table package. Internal data manipulation is now done using