+++ NEWS.md
@@ -1,4 +1,4 @@
-# treasury (development version)
+# treasury 0.4.0
 
 * Fixed bug in checking for {readxl} installation.
 * Improved documentation.